Refactor prototype code into separate files
