+++ Reqnroll.StepBindingSourceGenerator/BindingMethod.cs
@@ -1,0 +1,8 @@
+﻿namespace Reqnroll.StepBindingSourceGenerator;
+
+internal enum BindingMethod
+{
+    RegularExpression,
+    CucumberExpression,
+    MethodName
+}
