+++ Reqnroll.StepBindingSourceGenerator/CSharpStepBindingGenerator.cs
@@ -1,7 +1,4 @@
 ﻿using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.Text;
-using System.Collections;
 using System.Collections.Immutable;
 
 namespace Reqnroll.StepBindingSourceGenerator;
@@ -146,241 +143,4 @@ public class CSharpStepBindingGenerator : IIncrementalGenerator
                 emitter.EmitStepDefinitionCatalogClass(stepDefinitionClass));
         });
     }
-}
-
-/// <summary>
-/// Represents the pipline-relevant information about the declaration of a step definition.
-/// </summary>
-/// <param name="MethodDeclarationSyntax">The method declaration syntax which has been identified to be a step 
-/// definition.</param>
-internal record StepDefinitionSyntaxInfo(
-    MethodDeclarationSyntax MethodDeclarationSyntax,
-    MethodInfo Method,
-    StepKeywordMatch MatchedKeywords,
-    string? TextPattern);
-
-internal record MethodInfo(
-    string Name,
-    QualifiedTypeName DeclaringClassName,
-    DiagnosticLocation IdentifierLocation,
-    bool ReturnsVoid,
-    bool ReturnsTask,
-    bool IsAsync,
-    ComparableArray<ParameterInfo> Parameters);
-
-internal record struct QualifiedTypeName(string Namespace, string Name);
-
-internal record ParameterInfo(string Name, string ParameterType);
-
-internal enum StepKeywordMatch
-{
-    Given,
-    When,
-    Then,
-    Any
-}
-
-internal enum MethodInvocationStyle
-{
-    Synchronous,
-    Asynchronous
-}
-
-internal record StepDefinitionInfo(
-    string Name,
-    string DisplayName,
-    MethodInfo Method,
-    StepKeywordMatch MatchesKeywords,
-    BindingMethod BindingMethod,
-    string? TextPattern,
-    MethodInvocationStyle InvocationStyle);
-
-internal record StepDefinitionCatalogInfo(
-    string HintName,
-    QualifiedTypeName ClassName,
-    ComparableArray<StepDefinitionInfo> StepDefinitions);
-
-internal static class ComparableArray
-{
-    public static ComparableArray<T> CreateRange<T>(IEnumerable<T> items) => new(ImmutableArray.CreateRange(items));
-
-    public static ComparableArray<T> Create<T>(T item) => new(ImmutableArray.Create(item));
-
-    public static ComparableArray<T> CreateRange<T>(ReadOnlySpan<T> items)
-    {
-        var builder = ImmutableArray.CreateBuilder<T>(items.Length);
-
-        foreach (var item in items)
-        {
-            builder.Add(item);
-        }
-
-        return new ComparableArray<T>(builder.ToImmutable());
-    }
-}
-
-internal readonly struct ComparableArray<T>(ImmutableArray<T> values) : IEquatable<ComparableArray<T>>, IEnumerable<T>
-{
-    private readonly ImmutableArray<T> _values = values;
-
-    public static ComparableArray<T> Empty { get; } = new ComparableArray<T>(ImmutableArray<T>.Empty);
-
-    public int Length => _values.Length;
-
-    public bool IsDefaultOrEmpty => _values.IsDefaultOrEmpty;
-
-    public override bool Equals(object obj)
-    {
-        if (obj is ComparableArray<T> other)
-        {
-            return Equals(other);
-        }
-
-        return false;
-    }
-
-    public bool Equals(ComparableArray<T> other)
-    {
-        if (_values.IsDefaultOrEmpty)
-        {
-            return other._values.IsDefaultOrEmpty;
-        }
-
-        if (other._values.IsDefaultOrEmpty)
-        {
-            return false;
-        }
-
-        // This comparison checks if the two immutable arrays point to the same underlying array object.
-        if (_values == other._values)
-        {
-            return true;
-        }
-
-        if (_values.Length != other._values.Length)
-        {
-            return false;
-        }
-
-        for (var i = 0; i < _values.Length; i++)
-        {
-            if (!Equals(_values[i], other._values[i]))
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    public override int GetHashCode()
-    {
-        if (_values.IsDefaultOrEmpty)
-        {
-            return 0;
-        }
-
-        return _values.GetSequenceHashCode();
-    }
-
-    public ImmutableArray<T>.Enumerator GetEnumerator() => 
-        (_values.IsDefault ? ImmutableArray<T>.Empty : _values).GetEnumerator();
-
-    IEnumerator<T> IEnumerable<T>.GetEnumerator() => 
-        ((IEnumerable<T>)(_values.IsDefault ? ImmutableArray<T>.Empty : _values)).GetEnumerator();
-
-    IEnumerator IEnumerable.GetEnumerator() => 
-        ((IEnumerable<T>)(_values.IsDefault ? ImmutableArray<T>.Empty : _values)).GetEnumerator();
-
-    public static implicit operator ComparableArray<T>(ImmutableArray<T> array) => new(array);
-}
-
-internal enum BindingMethod
-{
-    RegularExpression,
-    CucumberExpression,
-    MethodName
-}
-
-internal sealed record DiagnosticInfo(
-    DiagnosticDescriptor Descriptor,
-    DiagnosticLocation Location,
-    params object?[]? MessageArguments)
-{
-    public DiagnosticSeverity Severity => Descriptor.DefaultSeverity;
-
-    public Diagnostic CreateDiagnostic() => Diagnostic.Create(Descriptor, Location, MessageArguments);
-
-    public bool Equals(DiagnosticInfo? other)
-    {
-        if (other is null)
-        {
-            return false;
-        }
-
-        if (!(Descriptor.Equals(other.Descriptor) &&
-            Location.Equals(other.Location)))
-        {
-            return false;
-        }
-        
-        if (MessageArguments == null)
-        {
-            return other.MessageArguments == null;
-        }
-
-        if (other.MessageArguments == null)
-        {
-            return false;
-        }
-
-        return MessageArguments.SequenceEqual(other.MessageArguments);
-    }
-
-    public override int GetHashCode()
-    {
-        unchecked
-        {
-            var hash = 4037797;
-
-            hash *= 1864987 + Descriptor.GetHashCode();
-            hash *= 1864987 + Location.GetHashCode();
-            hash *= 1864987 + MessageArguments?.GetSequenceHashCode() ?? 0;
-
-            return hash;
-        }
-    }
-}
-
-internal record struct DiagnosticLocation(string FilePath, TextSpan TextSpan, LinePositionSpan LineSpan)
-{
-    public static DiagnosticLocation None { get; } = default;
-
-    public readonly Location? ToLocation() => FilePath == null ? null : Location.Create(FilePath, TextSpan, LineSpan);
-
-    public static DiagnosticLocation CreateFrom(SyntaxNode? node) => node == null ? None : CreateFrom(node.GetLocation());
-
-    public static DiagnosticLocation CreateFrom(Location location)
-    {
-        if (location.SourceTree == null)
-        {
-            throw new ArgumentException("Diagnostic location must be from a source location.", nameof(location));
-        }
-
-        return new DiagnosticLocation(location.SourceTree.FilePath, location.SourceSpan, location.GetLineSpan().Span);
-    }
-
-    public static DiagnosticLocation CreateFrom(SyntaxReference? syntaxReference)
-    { 
-        if (syntaxReference == null)
-        {
-            return None;
-        }
-
-        return CreateFrom(syntaxReference.SyntaxTree.GetLocation(syntaxReference.Span));
-    }
-
-    public static DiagnosticLocation CreateFrom(SyntaxToken token) => CreateFrom(token.GetLocation());
-
-    public static implicit operator Location?(DiagnosticLocation location) => location.ToLocation();
 }
