+++ Reqnroll.StepBindingSourceGenerator/MethodInfo.cs
@@ -1,0 +1,9 @@
+﻿namespace Reqnroll.StepBindingSourceGenerator;
+
+internal record MethodInfo(
+    string Name,
+    QualifiedTypeName DeclaringClassName,
+    bool ReturnsVoid,
+    bool ReturnsTask,
+    bool IsAsync,
+    ComparableArray<ParameterInfo> Parameters);
