+++ Reqnroll.StepBindingSourceGenerator/MethodInvocationStyle.cs
@@ -1,0 +1,7 @@
+﻿namespace Reqnroll.StepBindingSourceGenerator;
+
+internal enum MethodInvocationStyle
+{
+    Synchronous,
+    Asynchronous
+}
