+++ Reqnroll.StepBindingSourceGenerator/ParameterInfo.cs
@@ -1,0 +1,3 @@
+﻿namespace Reqnroll.StepBindingSourceGenerator;
+
+internal record ParameterInfo(string Name, string ParameterType);
