+++ Reqnroll.StepBindingSourceGenerator/QualifiedTypeName.cs
@@ -1,0 +1,3 @@
+﻿namespace Reqnroll.StepBindingSourceGenerator;
+
+internal record struct QualifiedTypeName(string Namespace, string Name);
