+++ Reqnroll.StepBindingSourceGenerator/StepDefinitionCatalogInfo.cs
@@ -1,0 +1,6 @@
+﻿namespace Reqnroll.StepBindingSourceGenerator;
+
+internal record StepDefinitionCatalogInfo(
+    string HintName,
+    QualifiedTypeName ClassName,
+    ComparableArray<StepDefinitionInfo> StepDefinitions);
