+++ Reqnroll.StepBindingSourceGenerator/StepKeywordMatch.cs
@@ -1,0 +1,9 @@
+﻿namespace Reqnroll.StepBindingSourceGenerator;
+
+internal enum StepKeywordMatch
+{
+    Given,
+    When,
+    Then,
+    Any
+}
